+++ test/api/resolveConfig/webpack.config1.cjs
