+++ test/api/resolveConfig/webpack.config2.cjs
